SRTM I2C Service: Make buffer length configurable
This change allows the SRTM_I2C_BUF_LEN definition to be configured
outside of the library itself. This was necessary to avoid certain
quirks and limitations of the RPMsg I2C driver in Linux not being
able to send messages greater than the default of 16 bytes, which is
necessary for some I2C devices.