i.MX8M Plus SoC M7: Add support for FlexSPI, ECSPI, and cache
This change pulls in support for the MCUX ECSPI driver when the
appropriate config is enabled (CONFIG_SPI_MCUX_ECSPI).

Also, there were missing defines related to the FlexSPI peripheral
of the M7 core of the i.MX8M Plus Soc ("MIMX8ML8" target). This
change updates the main system and features header files to work
appropriately with the fsl_flexspi driver.

Finally, this change pulls in support for the MCUX cache driver
when the appropriate config is enabled (CONFIG_HAS_MCUX_CACHE) for
the M7 core of the i.MX8M Plus SoC ("MIMX8ML8" target).